backend/requirements/test.txt
# Test requirements for NexusPM Enterprise
# Professional testing stack for enterprise-grade quality assurance

# Core testing framework
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0              # Coverage reporting
pytest-xdist==3.3.1            # Parallel test execution
pytest-mock==3.12.0            # Advanced mocking capabilities

# Test data generation
factory-boy==3.3.0             # Model instance factories
Faker==19.12.0                 # Realistic fake data generation

# Database testing
pytest-factoryboy==2.6.0       # Factory Boy integration with pytest

# API testing
responses==0.24.1              # HTTP response mocking

# Performance and load testing
pytest-benchmark==4.0.0       # Performance benchmarking

# Code quality and linting integration
pytest-flake8==1.1.1          # Code style testing

# Reporting and documentation
pytest-html==4.1.1            # HTML test reports
pytest-json-report==1.5.0     # JSON test reports for CI/CD

# Additional utilities for enterprise testing
freezegun==1.2.2              # Time mocking for date/time tests
model-bakery==1.17.0          # Alternative to Factory Boy
django-test-plus==2.2.2       # Enhanced Django testing utilities   